move event classes to separate package
