+++ code/app/src/main/java/com/example/experiment_automata/backend/events/OnEventListener.java
@@ -1,4 +1,4 @@
-package com.example.experiment_automata.backend.experiments;
+package com.example.experiment_automata.backend.events;
 
 public interface OnEventListener {
     public void onEvent();
