+++ code/app/src/main/java/com/example/experiment_automata/backend/experiments/ExperimentManager.java
@@ -1,13 +1,14 @@
 package com.example.experiment_automata.backend.experiments;
 
 import android.location.Location;
-import android.media.MediaDrm;
 import android.util.Log;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import com.example.experiment_automata.backend.DataBase;
+import com.example.experiment_automata.backend.events.OnEventListener;
+import com.example.experiment_automata.backend.events.UpdateEvent;
 import com.example.experiment_automata.backend.trials.BinomialTrial;
 import com.example.experiment_automata.backend.trials.CountTrial;
 import com.example.experiment_automata.backend.trials.MeasurementTrial;
